+++ packages/block-editor/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### New Features
+
+-   Added `useBlockAttribute` hook to simplify handling block attributes when editing blocks ([#42391](https://github.com/WordPress/gutenberg/pull/42391)).
+
 ## 9.5.0 (2022-07-13)
 
 ## 9.4.0 (2022-06-29)